More changes in itrie
